# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipes API",
        "description": "API para gerenciamento de receitas culinárias",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all recipes with filters",
                "description": "Display a listing of all recipes.",
                "operationId": "8b9e556d4d747bd60232961951494bfb",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "servings_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "exact"
                    },
                    {
                        "name": "servings_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 4
                    },
                    {
                        "name": "prep_time_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "below"
                    },
                    {
                        "name": "prep_time_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 30
                    },
                    {
                        "name": "rating_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "above"
                    },
                    {
                        "name": "rating_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        },
                        "example": 4
                    },
                    {
                        "name": "comments_operator",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "exact",
                                "above",
                                "below"
                            ]
                        },
                        "example": "above"
                    },
                    {
                        "name": "comments_value",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 5
                    },
                    {
                        "name": "my_recipes",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": false
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "bolo"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Recipe"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 150
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "description": "Store a newly created recipe.",
                "operationId": "46448289940e329a644b891be9c66943",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "instructions"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Bolo de Chocolate",
                                        "nullable": true
                                    },
                                    "prep_time_minutes": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 30,
                                        "nullable": true
                                    },
                                    "servings": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 8,
                                        "nullable": true
                                    },
                                    "image": {
                                        "description": "Image file (max 5MB, formats: jpeg, jpg, png, gif, webp)",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "maxLength": 500,
                                        "example": "https://example.com/image.jpg",
                                        "nullable": true
                                    },
                                    "instructions": {
                                        "type": "string",
                                        "example": "Misture todos os ingredientes..."
                                    },
                                    "ingredients": {
                                        "type": "string",
                                        "example": "2 xícaras de farinha, 3 ovos...",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Recipe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Recipe": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sobremesas"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Bolo de Chocolate",
                        "nullable": true
                    },
                    "prep_time_minutes": {
                        "type": "integer",
                        "example": 30,
                        "nullable": true
                    },
                    "servings": {
                        "type": "integer",
                        "example": 8,
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "instructions": {
                        "type": "string",
                        "example": "Misture todos os ingredientes..."
                    },
                    "ingredients": {
                        "type": "string",
                        "example": "2 xícaras de farinha, 3 ovos...",
                        "nullable": true
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "comment": {
                                    "type": "string",
                                    "example": "Delicioso!"
                                },
                                "user": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        }
                                    },
                                    "type": "object"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "comments_count": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "ratings_count": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "average_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}